move array and multi parameter tests to their own files
